+++ app/frontend/src/App.jsx
@@ -8,6 +8,8 @@ import {StylesProvider, createMuiTheme} from '@material-ui/core/styles';
 import Snackbar from '@material-ui/core/Snackbar';
 import MuiAlert from '@material-ui/lab/Alert';
 import AddToHomescreen from 'react-add-to-homescreen';
+import useErrorBoundary from "use-error-boundary"
+
 
 import Menu from '@/Menu';
 import Entrance from '@/Entrance';
@@ -23,11 +25,16 @@ import PrivateRoute from '~/components/Menu/PrivateRoute';
 import store from './store';
 import {fetchAllowedUsers, fetchSubjects} from '~/actions/users';
 import {fetchMyToday} from '~/actions/calendar';
-import {fetchReasons, popNotification, updateOnline} from '~/actions/general';
+import {
+  fetchReasons,
+  newNotification,
+  popNotification,
+  updateOnline
+} from '~/actions/general';
 import {fetchMadors} from '~/actions/madors';
 
 const Alert = (props) => <MuiAlert elevation={6}
-  variant="filled" {...props} />;
+                                   variant="filled" {...props} />;
 
 
 export const App = (props) => {
@@ -90,7 +97,7 @@ export const App = (props) => {
       <Route path="/entrance" component={Entrance}/>
       <Route path="/" render={() => (
         <Drawer onDrag={onDrawerDrag} onToggle={onDrawerToggle}
-          onDragEnd={onDrawerDragEnd}>
+                onDragEnd={onDrawerDragEnd}>
           <DrawerMenu>
             <Menu avatar={avatar} avatarRef={avatarRef}/>
           </DrawerMenu>
@@ -112,16 +119,16 @@ export const App = (props) => {
             )}>
             <Switch>
               <PrivateRoute path="/hierarchy" component={Hierarchy}
-                key={reloadCount}
-                allowedPermissions={['admin', 'reporter']}/>
+                            key={reloadCount}
+                            allowedPermissions={['admin', 'reporter']}/>
               <PrivateRoute path="/operator" component={Operator}
-                key={reloadCount}
-                allowedPermissions={['admin', 'reporter']}/>
+                            key={reloadCount}
+                            allowedPermissions={['admin', 'reporter']}/>
               <PrivateRoute path="/commander" component={Commander}
-                key={reloadCount}
-                allowedPermissions={['admin', 'commander']}/>
+                            key={reloadCount}
+                            allowedPermissions={['admin', 'commander']}/>
               <PrivateRoute path="/" component={Dashboard}
-                key={reloadCount}/>
+                            key={reloadCount}/>
             </Switch>
           </DrawerContent>
         </Drawer>
@@ -138,6 +145,20 @@ export const StyledApp = (props) => {
   const notifications = useSelector((state) => state.general.notifications);
   const [notification, setNotification] = useState(null);
   const [showNotification, setShowNotification] = useState(false);
+  const {
+    ErrorBoundary, // class - The react component to wrap your children in. This WILL NOT CHANGE
+    didCatch, // boolean - Whether the ErrorBoundary catched something
+    error, // null or the error
+    errorInfo // null or the error info as described in the react docs
+  } = useErrorBoundary();
+
+  useEffect(() => {
+    if (didCatch) {
+      dispatch(newNotification({
+        message: 'Critical Error Occurred'
+      }))
+    }
+  }, [didCatch]);
 
   useEffect(() => {
     if (notifications.length === 0) return;
@@ -161,12 +182,27 @@ export const StyledApp = (props) => {
       <MUIThemeProvider theme={createMuiTheme(theme)}>
         <GlobalStyle/>
         <ThemeProvider theme={theme}>
-          <App/>
+          {
+            didCatch ?
+              <div style={{overflow: 'scroll'}}>
+                <h1>An error was occurred</h1>
+                Please report to devs:
+                <details style={{whiteSpace: 'pre-wrap'}}>
+                  {error.toString()}
+                  <br/>
+                  {errorInfo.componentStack}
+                </details>
+              </div>
+              :
+              <ErrorBoundary>
+                <App/>
+              </ErrorBoundary>
+          }
           <Snackbar open={showNotification}
-            autoHideDuration={notification?.timeout ?? SNACKBAR_TIMEOUT}
-            onClose={handleClose}>
+                    autoHideDuration={notification?.timeout ?? SNACKBAR_TIMEOUT}
+                    onClose={handleClose}>
             <Alert onClose={handleClose}
-              severity={notification?.severity ?? 'error'}>
+                   severity={notification?.severity ?? 'error'}>
               {notification?.message}
             </Alert>
           </Snackbar>