+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added more icons and cosmetics
 - Added success message on hierarchy page
 - Fixed notification hidden visual bug
+- No whitescreen on error (display traceback)
 
 ### [0.2.1-beta](https://github.com/TeamTash/one_report/compare/0.2.0-alpha...0.2.1-beta) - 21/02/2020
 - fixed recursion in hierarchy